+++ wasm/src/org.graalvm.wasm/src/org/graalvm/wasm/WasmOptions.java
@@ -78,4 +78,8 @@ public class WasmOptions {
     public static final String SATURATING_FLOAT_TO_INT_NAME = "saturating-float-to-int";
     @Option(name = SATURATING_FLOAT_TO_INT_NAME, help = "Use saturating-float-to-int conversion", category = OptionCategory.EXPERT, stability = OptionStability.EXPERIMENTAL) //
     public static final OptionKey<Boolean> SATURATING_FLOAT_TO_INT = new OptionKey<>(false);
+
+    public static final String SIGN_EXTENSION_OPS_NAME = "sign-extension-ops";
+    @Option(name = SIGN_EXTENSION_OPS_NAME, help = "Use sign-extension operators", category = OptionCategory.EXPERT, stability = OptionStability.EXPERIMENTAL) //
+    public static final OptionKey<Boolean> SIGN_EXTENSION_OPS = new OptionKey<>(false);
 }